Move general QMOperators one level up
