+++ src/qmoperators/QMDerivative.h
@@ -25,45 +25,26 @@
 
 #pragma once
 
-#include "qmfunctions/QMFunction.h"
-#include "qmoperators/QMOperator.h"
+#include <memory>
 
-/** @class QMPotential
- *
- * @brief Operator defining a multiplicative potential
- *
- * Inherits the general features of a complex function from QMFunction and
- * implements the multiplication of this function with an Orbital. The actual
- * function representing the operator needs to be implemented in the derived
- * classes, where the *re and *im FunctionTree pointers should be assigned in
- * the setup() function and deallocated in the clear() function.
- *
- */
+#include "QMOperator.h"
 
 namespace mrchem {
 
-class QMPotential : public QMFunction, public QMOperator {
+class QMDerivative final : public QMOperator {
 public:
-    explicit QMPotential(int adap, bool shared = false);
-    QMPotential(const QMPotential &pot) = delete;
-    QMPotential &operator=(const QMPotential &pot) = delete;
-    virtual ~QMPotential() = default;
+    QMDerivative(int d, std::shared_ptr<mrcpp::DerivativeOperator<3>> D, bool im = false);
 
-protected:
-    int adap_build;
+private:
+    const bool imag;     // add imaginary unit prefactor, for faster application
+    const int apply_dir; // Cartesian direction of derivative
+    std::shared_ptr<mrcpp::DerivativeOperator<3>> derivative;
 
-    ComplexDouble evalf(const mrcpp::Coord<3> &r) const override {
-        ComplexDouble out(0.0, 0.0), i(0.0, 1.0);
-        if (this->hasReal()) out += this->real().evalf(r);
-        if (this->hasImag()) out += i * this->imag().evalf(r);
-        return out;
-    }
+    bool isReal() const { return not(imag); }
+    bool isImag() const { return imag; }
 
     Orbital apply(Orbital inp) override;
     Orbital dagger(Orbital inp) override;
-
-    void calcRealPart(Orbital &out, Orbital &inp, bool dagger);
-    void calcImagPart(Orbital &out, Orbital &inp, bool dagger);
 };
 
 } // namespace mrchem
