+++ src/qmoperators/QMOperator.h
@@ -25,9 +25,8 @@
 
 #pragma once
 
-#include "MRCPP/Printer"
+#include <MRCPP/Printer>
 
-#include "chemistry/chemistry_fwd.h"
 #include "mrchem.h"
 #include "qmoperator_fwd.h"
