+++ src/qmoperators/one_electron/CMakeLists.txt
@@ -1,8 +1,4 @@
 target_sources(mrchem PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/KineticOperator.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/NuclearOperator.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/QMDerivative.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/QMIdentity.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/QMPotential.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/QMSpin.cpp
     )
