+++ src/qmoperators/one_electron/H_BB_dia.h
@@ -25,8 +25,9 @@
 
 #pragma once
 
-#include "PositionOperator.h"
 #include "qmoperators/RankTwoTensorOperator.h"
+
+#include "PositionOperator.h"
 
 namespace mrchem {
