+++ src/qmoperators/one_electron/H_E_quad.h
@@ -25,8 +25,9 @@
 
 #pragma once
 
-#include "PositionOperator.h"
 #include "qmoperators/RankTwoTensorOperator.h"
+
+#include "PositionOperator.h"
 
 /** Class H_E_quad
  *
