+++ src/qmoperators/two_electron/ExchangePotential.h
@@ -27,9 +27,10 @@
 
 #include <memory>
 
+#include "qmoperators/QMOperator.h"
+
 #include "qmfunctions/Orbital.h"
 #include "qmfunctions/qmfunction_fwd.h"
-#include "qmoperators/QMOperator.h"
 #include "utils/Bank.h"
 
 namespace mrchem {
