+++ src/qmoperators/two_electron/ExchangePotentialD1.h
@@ -28,8 +28,8 @@
 #include <memory>
 
 #include "ExchangePotential.h"
+
 #include "qmfunctions/qmfunction_fwd.h"
-#include "qmoperators/QMOperator.h"
 
 namespace mrchem {
