+++ src/qmoperators/two_electron/XCPotential.h
@@ -25,8 +25,9 @@
 
 #pragma once
 
+#include "qmoperators/QMPotential.h"
+
 #include "mrdft/MRDFT.h"
-#include "qmoperators/one_electron/QMPotential.h"
 
 /**
  * @class XCPotential
