+++ src/qmoperators/two_electron/XCPotentialD2.h
@@ -48,8 +48,6 @@
  * the XC potentials: either with explicit derivatives or gamma-type derivatives.
  */
 
-#include "MRCPP/MWFunctions"
-
 namespace mrchem {
 
 class XCPotentialD2 final : public XCPotential {
